Added Sacred artifact handling for checkpoints and TF logs
Eliminated logdir and ckptdir arguments. Instead, a temporary dir is now created each run in which the latest ONE checkpoint is stored as well as a single log from TF containing the graph. These are then sent as artifacts to Sacred at the end of the run and stored in the database.